# luasettings: backup previous file and auto-recover