Similar to what was done for docsettings in #3093 (will apply to settings.reader.lua, but also to plugins' own settings).
Had it since after my early MuPDF 1.12 kobo crashes, and it has recovered once when I had wrongly manually edited settings.reader.lua.