@poire-z 2019-03-15
I see on mobileread people having lost their settings.reader.lua on crash.
This PR was supposed to help on that - the previous is saved as .old - and recovered if the main one is corrupted.
And I guess it does help - cause i had no loss for over a year.

But it happened to me last week (cpuloop, before jit.opt tweaks) that both .lua and .lua.old must have been screwed. Possibly because I tried to suspend and resume and suspend, so both were written in a small time span and didn't get flushed to disk.
I don't know if we can do fsync() from lua - or force the kernel to write and sync without delay (would it hurt general usage on SDCards ?)
Or may be we shouldn't recycle the .lua to .lua.old if the .lua has not aged enough. Assuming that the previous .lua is say 3mn old, it must have been flushed to disk - and we can rename it to .lua.old.
Or may be it's just these renames and the directory entries list that get screwed up - and both the settings files would not be found anyway.

Better lose 3mn of settings changes than a lifetime of them :)
What do you think?