@NiLuJe 2019-03-16
FAT32 not being journaled, I'm not sure there's actually a silver bullet.

(Note that the reports on MR are somewhat quirkier, because the situation is much more complex on Kindle, where the filesystem we interact with is actually a fuse proxy. Inconsistencies between the fuse proxy and the underlying FS are easy to reproduce, especially when overwriting an existing file).