@NiLuJe 2019-03-16
In any case, much like we use & abuse `sync` in startup scripts, that's also exported by the glibc, so it'd be trivial to FFI. Might not help much, though.