@poire-z 2019-12-04
Again in #5663.
I'm not sure I understand what's specific about FAT32.
But with linux on ext3+, the content we write might be flushed to disk after 10/20s, while the open/rename of the file in the inode directory is remembered in the journal.
So, if we happen to flush/saveSettings 2 times in these 10s, and have a hard crash/hang, there will be no content to reader after the reboot, as there will be 2 files of 0 size.
https://github.com/koreader/koreader/blob/e8f91ab3997adbdfdbc645be46d51c62b8c8013e/frontend/luasettings.lua#L180-L195

So, on ext2+, a solution would be to use fsync, or to just not touch/rotate that settings to .old if the file we are moving is not older than 1 minute (so we know it has been flushed/synced by Linux) - by just adding a timestamp check on line 183:
- if settings.reader.lua has been modified less than 1 mn ago, keep writting to the same settings.reader.lua - keep the previous older .old as-is
- it settings.reader.lua has not been saved less than 1mn ago, move it to settings.reader.lua.old, and create a new settings.reader.lua with the content

With FAT32, if the diretory items rename is not journalized and if not remembered in case of hang, we might still find the older settings.reader.lua.old.
On ext3, we'll also still get a valid settings.reader.lua.old.

https://github.com/koreader/koreader/blob/e8f91ab3997adbdfdbc645be46d51c62b8c8013e/frontend/luasettings.lua#L27-L41

What do you think?