@poire-z 2019-12-04
Or, `os.execute("sync")` at the end of this `LuaSettings:flush()` :)