@Frenzie 2019-12-04
> I'm not sure I understand what's specific about FAT32.

Nothing, except as you mentioned it's one of the only non-journaling filesystem still in active use. Add a weird FUSE layer like Kindle apparently does and you've got a recipe for data corruption. With something like ext2/3/4 or NTFS the amount of potential issues is vastly reduced.

> a solution would be to use fsync

Does that do anything in practice that `:flush()` and/or `:close()` don't?