@poire-z 2019-12-04
Our :flush() / :close() only do the open()/write()/close() syscalls.
The fsync syscall blocks and only returns when it's written to disk.

```
DESCRIPTION
       fsync() transfers ("flushes") all modified in-core data of (i.e., modi-
       fied  buffer cache pages for) the file referred to by the file descrip-
       tor fd to the disk device (or other permanent storage device)  so  that
       all  changed information can be retrieved even if the system crashes or
       is rebooted.  This includes writing through or flushing a disk cache if
       present.   The  call  blocks until the device reports that the transfer
       has completed.

       As well as flushing the file data, fsync() also  flushes  the  metadata
       information associated with the file (see inode(7)).

       Calling  fsync()  does not necessarily ensure that the entry in the di-
       rectory containing the file has also reached disk.   For  that  an  ex-
       plicit fsync() on a file descriptor for the directory is also needed.
```
Otherwise, Linux delays the flush of what we do with open/write/close, and that can be tuned with:
https://www.kernel.org/doc/Documentation/sysctl/vm.txt
```
dirty_writeback_centisecs

The kernel flusher threads will periodically wake up and write `old' data
out to disk.  This tunable expresses the interval between those wakeups, in
100'ths of a second.
```
(at least, that's what I played with +10 years ago when I wanted to prevent bzzzzz the night when Linux was flushing its own stuff to some noisy laptop rotating hard drive :)