@NiLuJe 2019-12-04
The last of bit of the fsync ma page (fsync'ing the directory) is key, and may be non-trivial to actually achieve.

I'm not sure it'll help with FAT32 being a giant POS, but it'll probably be better ;). 

I don't have the code on hand, but making sure to use atomic operations on both sides would help, too: write the new config to a temp file, rename current to backup, rename temp to current. Note that rename cannot cross FS boundaries, so that temp file might need an extra copy to the same FS first, or being created there in the first place.