@poire-z 2019-12-09
OK, with https://github.com/koreader/koreader-base/pull/1021, may be we can go with:
```diff
 --- Writes settings to disk.
 function LuaSettings:flush()
     if not self.file then return end
+    local directory_updated = false
     if lfs.attributes(self.file, "mode") == "file" then
-        os.rename(self.file, self.file .. ".old")
+        -- As an additional safety measure (to the ffiutil.fsync* calls
+        -- used below), we only backup the file to .old when it has
+        -- not been modified in the last 60 seconds. This should ensure
+        -- in the case the fsync calls are not supported that the OS
+        -- may have itself sync'ed that file content in the meantime.
+        local mtime = lfs.attributes(self.file, "modification")
+        if mtime < os.time() - 60 then
+            os.rename(self.file, self.file .. ".old")
+            directory_updated = true -- fsync directory content too below
+        end
     end
     local f_out = io.open(self.file, "w")
     if f_out ~= nil then
         os.setlocale('C', 'numeric')
         f_out:write("-- we can read Lua syntax here!\nreturn ")
         f_out:write(dump(self.data))
         f_out:write("\n")
+        ffiutil.fsyncOpenedFile(f_out) -- force flush to the storage device
         f_out:close()
     end
+    if directory_updated then
+        -- Ensure the file renaming is flushed to storage device
+        ffiutil.fsyncDirectory(self.file)
+    end
     return self
 end
```