@poire-z 2019-12-09
I also tested this on my Android phone: got initially a feeling that switching books was slower - but reverting and back again and reverting - it was not coherent (slower sometimes without the fsync... and snappy with them... so... Android...)
Some timing of these fsync on my android 6 phone shows each fsync takes from 7ms to 15ms.
So, when quick switching books, with the 3 (+2 if rename to .old) fsync, that would make 25ms to 50ms added time blocking on these syscalls. (which always feels a bit sad for something 99.99% useless :)

Also, I feel that on Android it might not really be needed, there's less chance that it hangs or crashs, and we can recover finely if koreader is stuck without needing a hard reboot. So, should these be conditionnal. Or don't bother?